svg_loader: just renamed file name to simplify.
XmlParser without "Simple" is enough to address itself.
