+++ src/loaders/svg/tvgXmlParser.h
